删除评论

+++ app/Http/Controllers/RepliesController.php
@@ -33,14 +33,15 @@ class RepliesController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Reply  $reply
-     * @return \Illuminate\Http\Response
+     * @param Reply $reply
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(Reply $reply)
     {
-
+        $this->authorize('destroy', $reply);
         $reply->delete();
 
-        return redirect()->route('replies.index')->with('success', '评论删除成功');
+        return redirect()->to($reply->topic->link())->with('success', '评论删除成功！');
     }
 }
